requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
cffi==1.14.5
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
Django==3.2.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
